Refactored and organized TypeScript files! - Reorganized 'src/cli/index.ts' for better readability (- lines) - Added new functions/vars in 'src/types/index.ts' ( lines added,  file created) - Types have been consolidated and improved upon for a smoother experience - Cleaner code structure for our beloved Ollama CLI
